feat: make attendee/id api and add notes to order
